Windows CI: Unit Test move Unix specific struct field tests to _unix.go
